Filter used disk
This commit will do following:

1. Filter used disks.
2. Put variables in camel case.
3. Change maxPool and minPool integer type
   from int16 to int.
4. Ensure the min number of pool is at least created
   and not less than that.

+++ pkg/apis/openebs.io/v1alpha1/cas_pool.go
@@ -64,10 +64,10 @@ type CasPool struct {
 	PoolType string
 
 	// MaxPool is the maximum number of pool that should be provisioned
-	MaxPools int16
+	MaxPools int
 
 	// MinPool is the minimum number of pool that should be provisioned
-	MinPools int16
+	MinPools int
 
 	// Type is the CasPool type e.g. sparse or openebs-cstor
 	Type string
